// epms-payment-ui-develop/client/src/app/models/PgResponseData.ts
export class PgResponseData {
  decision?: string;
  pgmResponseCode?: string;
  pgmResponseMessage?: string;
  epmsTransactionID?: string;
  epmsOrderID?: string;
  programDefinedDate?: string;
  programSpecificKey?: string;
  methodOfPayment?: string;
  serviceCode?: string;
  resourceText?: string;
  paypalOrderId	?: string;
  originatingSystemCode?: string;
  oscSessionId?: string;
  baseCurrencyAmount?: string;
  maskedAccountNumber?: string;
  sourceType?: string;
  customerID?: string;
  transactionTime?: string;
  glProductText?: string;
  glProjectText?: string;
  glActivityText?: string;
  modeOfReceipt?: string;
  merchantDescriptor?: string;
  merchantDescriptorContact?: string;
  jeAccountNumber?: string;
  attemptCounter?: string;
  billingAddressCity?: string;
  billingAddressCompany?: string;
  billingAddressCountry?: string;
  billingAddressEmail?: string;
  billingAddressFirstName?: string;
  billingAddressLastName?: string;
  billingAddressPostalCode?: string;
  billingAddressState?: string;
  billingAddressLine1?: string;
  billingAddressLine2?: string;
  billingAddressPhoneNumber?: string;
  baseCurrency?: string;
  registrationDate?: string;
  customerType?: string;
  foreignAmount?: string;
  foreignCurrency?: string;
  storeID?: string;
  shippingAddressFirstName?: string;
  shippingAddressLastName?: string;
  shippingAddressLine1?: string;
  shippingAddressLine2?: string;
  shippingAddressPostalCode?: string;
  shippingAddressState?: string;
  shippingAddressCity?: string;
  shippingAddressCountry?: string;
  shippingAddressEmail?: string;
  shippingAdressCompany?: string;
  shippingAddressPhoneNumber?: string;
  displayLanguage?: string;
  merchantDefinedField1?: string;
  merchantDefinedField2?: string;
  merchantDefinedField3?: string;
  merchantDefinedField4?: string;
  merchantDefinedField5?: string;
  merchantDefinedField6?: string;
  merchantDefinedField7?: string;
  merchantDefinedField8?: string;
  merchantDefinedField9?: string;
  merchantDefinedField10?: string;
  testCenterCountry?: string;
  basetaxAmount?: string;
  exchangeRate?: string;
  inquiryRateId?: string;
  orderID?: string;
  numberOfPayments?: string;
  frequency?: string;
  subscriptionAmount?: string;
  orderAmount?: string;
  merchantErrorURL?: string;
  merchantRejectURL?: string;
  merchantSuccessURL?: string;
  merchantTimeoutURL?: string;
  merchantPostURL?: string;
  consentSigned?: string;
  referenceNumber?: string;
  authorizationCode?: string;
  expiryYear?: string;
  expiryMonth?: string;
  cardType?: string;
  cardNumberLast4?: string;
  payerID?: string;
  baseCurrencyCode?: string;
  accountType?: string;
  routingNumber?: string;
  transactionSignature?: string;
  signedFields?: string;
  timestamp?: string;
  ipgTransactionId?: string;
  modeOfAuthorization?: string;
  subscriptionTransactionId?: string;
  subscriptionRenew?: string;
  paymentTypeCode?: string;
  margin?: string;


  constructor() {
  }
}